% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/quality_control.R
\name{CalculateCutOff}
\alias{CalculateCutOff}
\alias{CalculateCutOff.default}
\alias{CalculateCutOff.Seurat}
\title{Calculate cut-off value}
\usage{
CalculateCutOff(object, ...)

\method{CalculateCutOff}{default}(
  object,
  metrics = c("mean", "median"),
  n.dev = 3,
  cut.off.end = c("paired", "lower", "higher"),
  cut.off.given = NULL,
  cut.off.given.only = FALSE,
  ...
)

\method{CalculateCutOff}{Seurat}(
  object,
  assay = NULL,
  feature = NULL,
  cells = NULL,
  slot = NULL,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{metrics}{Default is 'mean', and standard deviation will be used as discrete parameter. Median absolute deviation will be used for 'median' as metrics}

\item{n.dev}{Number for discrete parameter}

\item{cut.off.end}{Default is paired-end. Can set with 'lower' or 'higher' end}

\item{cut.off.given}{Add extra cut-off values}

\item{cut.off.given.only}{Whether or not to use the given cut-offs only, default is FALSE}

\item{assay}{Which assay to use}

\item{feature}{Which feature to calculate. Must be from either rownames of Seurat data or colnames of meta.data}

\item{cells}{Which cells to use. Default is all cells}

\item{slot}{Which slot to use. Default is counts slot}
}
\value{
Returns a vector storing cut-off(s) for outliars
}
\description{
Compute cut-off for outliars of an given feature.
}
