% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CellCycleEnrichScore}
\alias{CellCycleEnrichScore}
\title{Calculate cell-cycle z-score}
\usage{
CellCycleEnrichScore(
  object,
  assay = NULL,
  s.features = NULL,
  g2m.features = NULL,
  min.cells = 3,
  N = 100,
  seed = 42
)
}
\arguments{
\item{object}{A Seurat object}

\item{assay}{Which assay to use}

\item{s.features}{A vector of features associated with S phase}

\item{g2m.features}{A vector of features associated with G2M phase}

\item{min.cells}{Include features detected in at least this many cells}

\item{N}{Number of random background feature sets, default is 100}

\item{seed}{Set a random seed}
}
\value{
A Seurat object with the following columns added to object meta data: s.score, g2m.score, and phase
}
\description{
Another method to assign cell-cycle phase by computing z-score
}
\references{
Mayer, C., Hafemeister, C., Bandler, R. C., Machold, R., Batista Brito, R., Jaglin, X., … Satija, R. (2018). Developmental diversification of cortical inhibitory interneurons. Nature, 555(7697), 457–462. https://doi.org/10.1038/nature25999
}
