% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R, R/generics.R
\name{DBclustering.Seurat}
\alias{DBclustering.Seurat}
\alias{DBclustering.DimReduc}
\alias{DBclustering.default}
\alias{DBclustering}
\title{Perform spectral density clustering on single cells}
\usage{
\method{DBclustering}{Seurat}(
  object,
  assay = NULL,
  reduction = "tsne",
  key = NULL,
  dim.1 = 1,
  dim.2 = 2,
  eps,
  minPts = 5,
  seed.use = 42,
  ...
)

\method{DBclustering}{DimReduc}(
  object,
  key = NULL,
  dim.1 = 1,
  dim.2 = 2,
  eps,
  minPts = 5,
  seed.use = 42,
  ...
)

\method{DBclustering}{default}(object, eps, minPts = 5, seed.use = 42, ...)

DBclustering(object, ...)
}
\arguments{
\item{object}{An object}

\item{assay}{Assay to use in DBSCAN clustering}

\item{reduction}{Which dimensional reduction to use, default is tsne.}

\item{key}{Key used in provided DimReduc object}

\item{dim.1}{First dimension to use}

\item{dim.2}{second dimension to use}

\item{eps}{size of the epsilon neighborhood. Lower value to get more fine-scale clustering.}

\item{minPts}{number of minimum points in the eps region (for core points). Default is 5 points.}

\item{seed.use}{Random seed for the dbscan function}

\item{...}{Arguments passed to other methods and destiny}
}
\value{
Returns a Seurat object where the idents have been updated with new cluster info;
latest clustering results will be stored in object metadata under 'DB_clusters'.
Note that 'DB_clusters' will be overwritten everytime FindClusters is run
}
\description{
Find point clounds single cells in a two-dimensional space using density clustering (DBSCAN).
}
\examples{
pbmc_small
# Density based clustering on the first two tSNE dimensions
pbmc_small <- DBclustering(pbmc_small)

}
