% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{LabelClusters}
\alias{LabelClusters}
\title{Label clusters on a ggplot2-based scatter plot}
\usage{
LabelClusters(
  plot,
  id,
  clusters = NULL,
  labels = NULL,
  split.by = NULL,
  repel = TRUE,
  fill = "white",
  alpha.box = 0,
  alpha.text = 1,
  ...
)
}
\arguments{
\item{plot}{A ggplot2-based scatter plot}

\item{id}{Name of variable used for coloring scatter plot}

\item{clusters}{Vector of cluster ids to label}

\item{labels}{Custom labels for the clusters}

\item{split.by}{Split labels by some grouping label, useful when using
\code{\link[ggplot2]{facet_wrap}} or \code{\link[ggplot2]{facet_grid}}}

\item{repel}{Use \code{geom_label_repel} to create nicely-repelled labels}

\item{fill}{When use \code{geom_label_repel}, set the color of label boxes}

\item{alpha.box}{When use \code{geom_label_repel}, set the transparency (0-1) of label boxes}

\item{alpha.text}{When use \code{geom_label_repel}, set the transparency (0-1) of label texts}

\item{...}{Extra parameters to \code{\link[ggrepel]{geom_label_repel}} or \code{\link[ggplot2]{geom_text}}, such as \code{size}}
}
\value{
A ggplot2-based scatter plot with cluster labels
}
\description{
Label clusters on a ggplot2-based scatter plot
}
\examples{
plot <- DimPlot(object = pbmc_small)
LabelClusters(plot = plot, id = 'ident')

}
\seealso{
\code{\link[ggrepel]{geom_label_repel}} \code{\link[ggplot2]{geom_text}}
}
