% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{MultiFeaturePlot}
\alias{MultiFeaturePlot}
\title{Visualize 'features' on a dimensional reduction plot, with a modified type of legend}
\usage{
MultiFeaturePlot(
  object,
  features,
  ncol = NULL,
  check.italic = FALSE,
  do.bold = TRUE,
  plot.title.size = NULL,
  axis.type = c("default", "keep line", "keep title", "no axis"),
  line.size = 0.5,
  axis.title.size = NULL,
  axis.ticks.size = NULL,
  axis.text.size = NULL,
  show.legend = TRUE,
  legend.title = NULL,
  legend.title.size = NULL,
  legend.text.size = NULL,
  legend.key.size = 0.5,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{Vector of features to plot. Features can come from:
\itemize{
    \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
    \item A column name from meta.data (e.g. mitochondrial percentage - "percent.mito")
    \item A column name from a \code{DimReduc} object corresponding to the cell embedding values
    (e.g. the PC 1 scores - "PC_1")
}}

\item{ncol}{Number of columns to combine multiple feature plots to, ignored if split.by is not NULL}

\item{check.italic}{Whether or not to use italic text on gene names}

\item{do.bold}{Whether or not to use bold font on plot title}

\item{plot.title.size}{Plot title size}

\item{axis.type}{Select one of four axis types}

\item{line.size}{Size of axis line}

\item{axis.title.size, axis.ticks.size, axis.text.size}{Parameters about axis to be passed to \code{\link{theme}}}

\item{show.legend}{Whether or not to show legend}

\item{legend.title}{A user-defined title of legend}

\item{legend.title.size, legend.text.size, legend.key.size}{Parameters about legend to be passed to \code{\link{theme}}}

\item{...}{Extra parameters passed on to \code{\link{FeaturePlot}}}
}
\value{
A ggplot object
}
\description{
Colors single cells on a dimensional reduction plot according to a 'feature' (i.e. gene expression, PC scores, number of genes detected, etc.)
}
