% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/preprocess.R
\name{RegressOutNBreg}
\alias{RegressOutNBreg}
\alias{RegressOutNBreg.default}
\alias{RegressOutNBreg.Seurat}
\title{Regress out technical effects and cell cycle using regularized Negative Binomial regression (old Seurat function)}
\usage{
RegressOutNBreg(object, ...)

\method{RegressOutNBreg}{default}(
  object,
  latent.data,
  n.features = NULL,
  min.cells = 3,
  regress.features = NULL,
  res.clip.range = c(-30, 30),
  min.theta = 0.01,
  residual.type = "pearson",
  bin.size = 128,
  theta.given = NULL
)

\method{RegressOutNBreg}{Seurat}(
  object,
  assay = NULL,
  new.assay.name = NULL,
  cells = NULL,
  latent.var,
  slot = "scale.data",
  variable.features.n = 2000,
  variable.features.z.th = 1,
  return.variable.features = FALSE,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{latent.data}{effects to regress out (data.frame)}

\item{n.features}{number of features to use when estimating theta (default uses all features)}

\item{min.cells}{only use features that have been observed in at least this many cells}

\item{regress.features}{feature to run regression for (default uses all features)}

\item{res.clip.range}{numeric of length two specifying the min and max values the results will be clipped to}

\item{min.theta}{minimum theta to use in NB regression}

\item{residual.type}{string specifying the type of residual used (default is pearson)}

\item{bin.size}{number of features to put in each bin (to show progress)}

\item{theta.given}{skip the first step and use the fitted thetas from a previous run}

\item{assay}{Name of assay to pull the count data from}

\item{new.assay.name}{Name for the new assay containing the normolized data}

\item{cells}{Cells used to do regularized NB regression; default is NULL}

\item{latent.var}{The independent variables to regress out as a character vector; must match column names in meta.data}

\item{slot}{Slot where Pearson residual will return; default is scale.data}

\item{variable.features.n}{Use this many features as variable features after ranking by residual z-score; default is 2000}

\item{variable.features.z.th}{Instead of setting a fixed number of variable features, use this residual variance cutoff; this is only used when \code{variable.features.n} is set to NULL; default is 1}

\item{return.variable.features}{Whether or not return the variable features; default is FALSE}
}
\value{
Returns a Seurat object with pearson residuals. Intermediate results are saved in a misc slot
}
\description{
Remove unwanted effects from umi data and set scale.data to Pearson residuals
Uses future_lapply; you can set the number of cores it will use to n with plan(strategy = 'multicore', workers = n).
If n.features is set, only a (somewhat-random) subset of features is used for estimating the initial model parameters.
}
\references{
Mayer, C., Hafemeister, C., Bandler, R. C., Machold, R., Batista Brito, R., Jaglin, X., … Satija, R. (2018). Developmental diversification of cortical inhibitory interneurons. Nature, 555(7697), 457–462. https://doi.org/10.1038/nature25999
}
