% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RidgePlot}
\alias{RidgePlot}
\title{Single cell ridge plot (Modified version)}
\usage{
RidgePlot(
  object,
  features,
  cols = NULL,
  idents = NULL,
  sort = FALSE,
  assay = NULL,
  group.by = NULL,
  x.lab = NULL,
  y.lab = NULL,
  check.italic = FALSE,
  do.bold = TRUE,
  plot.title.size = NULL,
  line.size = 0.5,
  axis.title.size = NULL,
  axis.ticks.size = NULL,
  axis.text.size = NULL,
  y.max = NULL,
  same.y.lims = FALSE,
  log = FALSE,
  ncol = NULL,
  combine = c("CombinePlots", "patchwork", "None"),
  slot = "data",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot (gene expression, metrics, PC scores,
anything that can be retreived by FetchData)}

\item{cols}{Colors to use for plotting}

\item{idents}{Which classes to include in the plot (default is all)}

\item{sort}{Sort identity classes (on the x-axis) by the average
expression of the attribute being potted, can also pass 'increasing' or 'decreasing' to change sort direction}

\item{assay}{Name of assay to use, defaults to the active assay}

\item{group.by}{Group (color) cells in different ways (for example, orig.ident)}

\item{x.lab}{Title of x axis}

\item{y.lab}{Title of y axis}

\item{check.italic}{Whether or not to use italic text on gene names}

\item{do.bold}{Whether or not to use bold font on plot title}

\item{plot.title.size, axis.title.size, axis.ticks.size, axis.text.size}{Parameters about axis to be passed to \code{\link{theme}}}

\item{line.size}{Size of axis line}

\item{y.max}{Maximum y axis value}

\item{same.y.lims}{Set all the y-axis limits to the same values}

\item{log}{plot the feature axis on log scale}

\item{ncol}{Number of columns if multiple plots are displayed}

\item{combine}{Combine plots using cowplot::plot_grid or patchwork package. If set 'None', will return a list of ggplot objects.}

\item{slot}{Use non-normalized counts data for plotting}

\item{...}{Extra parameters passed on to \code{\link{CombinePlots}}}
}
\description{
Draws a ridge plot of single cell data (gene expression, metrics, PC
scores, etc.)
}
