% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{SCTransformWrapper}
\alias{SCTransformWrapper}
\title{Wrapper for SCTransform}
\usage{
SCTransformWrapper(
  object,
  assay = NULL,
  get.cellcycle.score = FALSE,
  min.cells = 5,
  min.cells.cc = NULL,
  s.features = NULL,
  g2m.features = NULL,
  bg.N = 100,
  seed = 42,
  vars.to.regress = NULL,
  variable.features.to.remove = NULL,
  future.multicore = TRUE,
  workers = NULL,
  ...
)
}
\arguments{
\item{object}{A seurat object}

\item{assay}{Name of assay to pull the count data from; default is 'RNA'}

\item{get.cellcycle.score}{Whether or not to calculate cell-cycle enrich score}

\item{min.cells}{Include features detected in at least this many cells}

\item{min.cells.cc}{Include features detected in at least this many cells when assign cell-cycle phase. If null, will be as same as min.cells}

\item{s.features}{A vector of features associated with S phase}

\item{g2m.features}{A vector of features associated with G2M phase}

\item{bg.N}{Number of random background feature sets, default is 100}

\item{seed}{Set a random seed}

\item{vars.to.regress}{Variables to regress out in a second non-regularized linear
regression. For example, percent.mito. Default is NULL}

\item{variable.features.to.remove}{A vector of features to be removed from variable features}

\item{future.multicore}{Whether or not to use multicore computation}

\item{workers}{Number of workers used for future.multicore}

\item{...}{Additional parameters passed to \code{\link{SCTransform}}}
}
\description{
Wrapper for SCTransform and CellCycleEnrichScore
}
